Clear halt on Bulk-IN endpoint during device open
It is normally not necessary to call clear_halt() on the Bulk-IN endpoint.
However some devices do not correctly reset the DATA0/DATA1 toggle
of the Bulk-IN endpoint during set_configuration(). These devices become
desynchronized after set_configuration(), causing data loss of the first
message sent by the device.

Explicitly calling clear_halt() on the Bulk-IN endpoint resets the
toggle state. This acts as a workaround for data toggle problems with
some devices (e.g. Thorlabs PM16-120).